fix(opentelemetry-instrumentation-asgi): Correct http.url attribute generation (#2477)
