+++ instrumentation/opentelemetry-instrumentation-starlette/tests/test_starlette_instrumentation.py
@@ -18,7 +18,7 @@ from unittest.mock import patch
 
 from starlette import applications
 from starlette.responses import PlainTextResponse
-from starlette.routing import Route
+from starlette.routing import Mount, Route
 from starlette.testclient import TestClient
 from starlette.websockets import WebSocket
 
@@ -101,6 +101,43 @@ class TestStarletteManualInstrumentation(TestBase):
             self.assertEqual(
                 span.instrumentation_scope.name,
                 "opentelemetry.instrumentation.starlette",
+            )
+
+    def test_sub_app_starlette_call(self):
+        """
+        This test is to ensure that a span in case of a sub app targeted contains the correct server url
+        """
+
+        self._client.get("/sub/home")
+        spans = self.memory_exporter.get_finished_spans()
+        self.assertEqual(len(spans), 3)
+        for span in spans:
+            # As we are only looking to the "outer" app, we would see only the "GET /sub" spans
+            self.assertIn("GET /sub", span.name)
+
+        # We now want to specifically test all spans including the
+        # - HTTP_TARGET
+        # - HTTP_URL
+        # attributes to be populated with the expected values
+        spans_with_http_attributes = [
+            span
+            for span in spans
+            if (
+                SpanAttributes.HTTP_URL in span.attributes
+                or SpanAttributes.HTTP_TARGET in span.attributes
+            )
+        ]
+
+        # expect only one span to have the attributes
+        self.assertEqual(1, len(spans_with_http_attributes))
+
+        for span in spans_with_http_attributes:
+            self.assertEqual(
+                "/sub/home", span.attributes[SpanAttributes.HTTP_TARGET]
+            )
+            self.assertEqual(
+                "http://testserver/sub/home",
+                span.attributes[SpanAttributes.HTTP_URL],
             )
 
     def test_starlette_route_attribute_added(self):
@@ -250,13 +287,20 @@ class TestStarletteManualInstrumentation(TestBase):
         def health(_):
             return PlainTextResponse("ok")
 
+        def sub_home(_):
+            return PlainTextResponse("sub hi")
+
+        sub_app = applications.Starlette(routes=[Route("/home", sub_home)])
+
         app = applications.Starlette(
             routes=[
                 Route("/foobar", home),
                 Route("/user/{username}", home),
                 Route("/healthzz", health),
-            ]
+                Mount("/sub", app=sub_app),
+            ],
         )
+
         return app
 
 
@@ -354,6 +398,72 @@ class TestAutoInstrumentation(TestStarletteManualInstrumentation):
         spans = self.memory_exporter.get_finished_spans()
         self.assertEqual(len(spans), 0)
 
+    def test_sub_app_starlette_call(self):
+        """
+        !!! Attention: we need to override this testcase for the auto-instrumented variant
+            The reason is, that with auto instrumentation, the sub app is instrumented as well
+            and therefore we would see the spans for the sub app as well
+
+        This test is to ensure that a span in case of a sub app targeted contains the correct server url
+        """
+
+        self._client.get("/sub/home")
+        spans = self.memory_exporter.get_finished_spans()
+        self.assertEqual(len(spans), 6)
+
+        for span in spans:
+            # As we are only looking to the "outer" app, we would see only the "GET /sub" spans
+            #   -> the outer app is not aware of the sub_apps internal routes
+            sub_in = "GET /sub" in span.name
+            # The sub app spans are named GET /home as from the sub app perspective the request targets /home
+            #   -> the sub app is technically not aware of the /sub prefix
+            home_in = "GET /home" in span.name
+
+            # We expect the spans to be either from the outer app or the sub app
+            self.assertTrue(
+                sub_in or home_in,
+                f"Span {span.name} does not have /sub or /home in its name",
+            )
+
+        # We now want to specifically test all spans including the
+        # - HTTP_TARGET
+        # - HTTP_URL
+        # attributes to be populated with the expected values
+        spans_with_http_attributes = [
+            span
+            for span in spans
+            if (
+                SpanAttributes.HTTP_URL in span.attributes
+                or SpanAttributes.HTTP_TARGET in span.attributes
+            )
+        ]
+
+        # We now expect spans with attributes from both the app and its sub app
+        self.assertEqual(2, len(spans_with_http_attributes))
+
+        # Due to a potential bug in starlettes handling of sub app mounts, we can
+        # check only the server kind spans for the correct attributes
+        # The internal one generated by the sub app is not yet producing the correct attributes
+        server_span = next(
+            (
+                span
+                for span in spans_with_http_attributes
+                if span.kind == SpanKind.SERVER
+            ),
+            None,
+        )
+
+        self.assertIsNotNone(server_span)
+        # As soon as the bug is fixed for starlette, we can iterate over spans_with_http_attributes here
+        # to verify the correctness of the attributes for the internal span as well
+        self.assertEqual(
+            "/sub/home", server_span.attributes[SpanAttributes.HTTP_TARGET]
+        )
+        self.assertEqual(
+            "http://testserver/sub/home",
+            server_span.attributes[SpanAttributes.HTTP_URL],
+        )
+
 
 class TestAutoInstrumentationHooks(TestStarletteManualInstrumentationHooks):
     """
@@ -373,6 +483,72 @@ class TestAutoInstrumentationHooks(TestStarletteManualInstrumentationHooks):
     def tearDown(self):
         self._instrumentor.uninstrument()
         super().tearDown()
+
+    def test_sub_app_starlette_call(self):
+        """
+        !!! Attention: we need to override this testcase for the auto-instrumented variant
+            The reason is, that with auto instrumentation, the sub app is instrumented as well
+            and therefore we would see the spans for the sub app as well
+
+        This test is to ensure that a span in case of a sub app targeted contains the correct server url
+        """
+
+        self._client.get("/sub/home")
+        spans = self.memory_exporter.get_finished_spans()
+        self.assertEqual(len(spans), 6)
+
+        for span in spans:
+            # As we are only looking to the "outer" app, we would see only the "GET /sub" spans
+            #   -> the outer app is not aware of the sub_apps internal routes
+            sub_in = "GET /sub" in span.name
+            # The sub app spans are named GET /home as from the sub app perspective the request targets /home
+            #   -> the sub app is technically not aware of the /sub prefix
+            home_in = "GET /home" in span.name
+
+            # We expect the spans to be either from the outer app or the sub app
+            self.assertTrue(
+                sub_in or home_in,
+                f"Span {span.name} does not have /sub or /home in its name",
+            )
+
+        # We now want to specifically test all spans including the
+        # - HTTP_TARGET
+        # - HTTP_URL
+        # attributes to be populated with the expected values
+        spans_with_http_attributes = [
+            span
+            for span in spans
+            if (
+                SpanAttributes.HTTP_URL in span.attributes
+                or SpanAttributes.HTTP_TARGET in span.attributes
+            )
+        ]
+
+        # We now expect spans with attributes from both the app and its sub app
+        self.assertEqual(2, len(spans_with_http_attributes))
+
+        # Due to a potential bug in starlettes handling of sub app mounts, we can
+        # check only the server kind spans for the correct attributes
+        # The internal one generated by the sub app is not yet producing the correct attributes
+        server_span = next(
+            (
+                span
+                for span in spans_with_http_attributes
+                if span.kind == SpanKind.SERVER
+            ),
+            None,
+        )
+
+        self.assertIsNotNone(server_span)
+        # As soon as the bug is fixed for starlette, we can iterate over spans_with_http_attributes here
+        # to verify the correctness of the attributes for the internal span as well
+        self.assertEqual(
+            "/sub/home", server_span.attributes[SpanAttributes.HTTP_TARGET]
+        )
+        self.assertEqual(
+            "http://testserver/sub/home",
+            server_span.attributes[SpanAttributes.HTTP_URL],
+        )
 
 
 class TestAutoInstrumentationLogic(unittest.TestCase):
